+++ index.js
@@ -1,39 +1,118 @@
+'use strict'
 /* global fetch */
 
-var baseUrl = 'https://www.cryptocompare.com/api/data/'
+const baseUrl = 'https://min-api.cryptocompare.com/data/'
 
 function fetchJSON (url) {
   return fetch(url)
     .then(res => res.json())
     .then(body => {
-      // 'response' is a CryptoCompare field
-      if (body.Response !== 'Success') throw new Error(body.Message)
-      return body.Data
+      if (body.Response === 'Error') throw body.Message
+      return body
     })
 }
 
-function coinList () {
-  var url = baseUrl + 'coinlist/'
+function price (fsym, tsyms, options) {
+  options = options || {}
+  let url = `${baseUrl}price?fsym=${fsym}&tsyms=${tsyms}`
+  if (options.exchanges) url += `&e=${options.exchanges}`
+  if (options.tryConversion === false) url += '&tryConversion=false'
   return fetchJSON(url)
 }
 
-function price (fsym, tsyms, useBtc) {
-  if (!Array.isArray(tsyms)) tsyms = [tsyms]
-  var url = baseUrl + 'price?fsym=' + fsym + '&tsyms=' + tsyms.join(',')
-  if (useBtc) url += 'usebtc=true'
+function priceMulti (fsyms, tsyms, options) {
+  options = options || {}
+  let url = `${baseUrl}pricemulti?fsyms=${fsyms}&tsyms=${tsyms}`
+  if (options.exchanges) url += `&e=${options.exchanges}`
+  if (options.tryConversion === false) url += '&tryConversion=false'
   return fetchJSON(url)
 }
 
-function priceHistorical (fsym, tsyms, time) {
-  if (!Array.isArray(tsyms)) tsyms = [tsyms]
-  if (!time instanceof Date) throw new Error('time parameter must be an instance of Date.')
-  time = Math.floor(time.getTime() / 1000)
-  var url = baseUrl + 'pricehistorical?fsym=' + fsym + '&tsyms=' + tsyms.join(',') + '&ts=' + time
-  return fetchJSON(url)
+function priceFull (fsyms, tsyms, options) {
+  options = options || {}
+  let url = `${baseUrl}pricemultifull?fsyms=${fsyms}&tsyms=${tsyms}`
+  if (options.exchanges) url += `&e=${options.exchanges}`
+  if (options.tryConversion === false) url += '&tryConversion=false'
+  // We want the RAW data, not the DISPLAY data:
+  return fetchJSON(url).then(result => result.RAW)
+}
+
+function priceHistorical (fsym, tsyms, time, options) {
+  options = options || {}
+  time = dateToTimestamp(time)
+  let url = `${baseUrl}pricehistorical?fsym=${fsym}&tsyms=${tsyms}&ts=${time}`
+  if (options.exchanges) url += `&e=${options.exchanges}`
+  if (options.tryConversion === false) url += '&tryConversion=false'
+  // The API returns json with an extra layer of nesting, so remove it
+  return fetchJSON(url).then(result => result[fsym])
+}
+
+function generateAvg (fsym, tsym, markets, tryConversion) {
+  let url = `${baseUrl}generateAvg?fsym=${fsym}&tsym=${tsym}&markets=${markets}`
+  if (tryConversion === false) url += '&tryConversion=false'
+  return fetchJSON(url).then(result => result.RAW)
+}
+
+function topPairs (fsym, limit) {
+  let url = `${baseUrl}top/pairs?fsym=${fsym}`
+  if (limit) url += `&limit=${limit}`
+  return fetchJSON(url).then(result => result.Data)
+}
+
+function topExchanges (fsym, tsym, limit) {
+  let url = `${baseUrl}top/exchanges?fsym=${fsym}&tsym=${tsym}`
+  if (limit) url += `&limit=${limit}`
+  return fetchJSON(url).then(result => result.Data)
+}
+
+function histoDay (fsym, tsym, options) {
+  options = options || {}
+  if (options.timestamp) options.timestamp = dateToTimestamp(options.timestamp)
+  let url = `${baseUrl}histoday?fsym=${fsym}&tsym=${tsym}`
+  if (options.limit === 'none') url += '&allData=true'
+  else if (options.limit) url += `&limit=${options.limit}`
+  if (options.tryConversion === false) url += '&tryConversion=false'
+  if (options.aggregate) url += `&aggregate=${options.aggregate}`
+  if (options.timestamp) url += `&toTs=${options.timestamp}`
+  return fetchJSON(url).then(result => result.Data)
+}
+
+function histoHour (fsym, tsym, options) {
+  options = options || {}
+  if (options.timestamp) options.timestamp = dateToTimestamp(options.timestamp)
+  let url = `${baseUrl}histohour?fsym=${fsym}&tsym=${tsym}`
+  if (options.limit) url += `&limit=${options.limit}`
+  if (options.tryConversion === false) url += '&tryConversion=false'
+  if (options.aggregate) url += `&aggregate=${options.aggregate}`
+  if (options.timestamp) url += `&toTs=${options.timestamp}`
+  return fetchJSON(url).then(result => result.Data)
+}
+
+function histoMinute (fsym, tsym, options) {
+  options = options || {}
+  if (options.timestamp) options.timestamp = dateToTimestamp(options.timestamp)
+  let url = `${baseUrl}histominute?fsym=${fsym}&tsym=${tsym}`
+  if (options.limit) url += `&limit=${options.limit}`
+  if (options.tryConversion === false) url += '&tryConversion=false'
+  if (options.aggregate) url += `&aggregate=${options.aggregate}`
+  if (options.timestamp) url += `&toTs=${options.timestamp}`
+  return fetchJSON(url).then(result => result.Data)
+}
+
+function dateToTimestamp (date) {
+  if (!(date instanceof Date)) throw new Error('timestamp must be an instance of Date.')
+  return Math.floor(date.getTime() / 1000)
 }
 
 module.exports = {
-  coinList,
   price,
-  priceHistorical
+  priceMulti,
+  priceFull,
+  priceHistorical,
+  generateAvg,
+  topPairs,
+  topExchanges,
+  histoDay,
+  histoHour,
+  histoMinute
 }