+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Unreleased
 ----------
 
+Switched to https://min-api.cryptocompare.com/
+
+- removed: `coinList()`
+- changed: `price()` returns a different data structure
+- changed: `priceHistorical()` returns a different data structure
+- changed: `price()` and `priceHistorical)()` function signature has changed
+- added: `priceMulti()`
+- added: `priceFull()`
+- added: `topPairs()`
+- added: `topExchanges()`
+- added: `histoDay()`
+- added: `histoHour()`
+- added: `histoMinute()`
+- added: `generateAvg()`
+- added: `exchanges` option for methods that support it
+
 0.0.2 / 2016-11-21
 ------------------
 - bug fix: `priceHistorical()` - wrong endpoint